nit: rename conversions file
